# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: API for user management including authentication and CRUD operations
  version: 1.0.0

servers:
  - url: http://localhost:5000

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
              required:
                - username
                - password
                - email
      responses:
        '201':
          description: User created successfully
        '400':
          description: Missing username or password, or user already exists

  /auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Bad username or password

  /auth/reset_password:
    post:
      summary: Reset user password (requires authentication)
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
              required:
                - old_password
                - new_password
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Invalid password

  /auth/forget_password:
    post:
      summary: Request password reset link
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: Password reset link generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  reset_link:
                    type: string
        '404':
          description: Email not found

  /auth/reset_password/{token}:
    post:
      summary: Reset password using token
      tags:
        - Authentication
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_password:
                  type: string
              required:
                - new_password
      responses:
        '200':
          description: Password has been updated
        '400':
          description: Invalid or expired token, or new password is required
        '404':
          description: User not found

  /user/{user_id}:
    put:
      summary: Update user information (Admin only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                active:
                  type: boolean
                role:
                  type: string
                  enum: [USER, ADMIN]
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Invalid role provided
        '403':
          description: Not authorized to modify this user
        '404':
          description: User not found
    delete:
      summary: Delete a user (Admin only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '403':
          description: Not authorized to delete this user
        '404':
          description: User not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT